+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Remove parenthesis from assay type.
 - Assume Latin-1 encoding for TSVs rather than UTF-8.
 - Update LC-MS fields.
+- Separate level-2 schemas in source.
 
 ## [v0.0.2](https://github.com/hubmapconsortium/ingest-validation-tools/tree/v0.0.2) - 2020-04-25
 ### Added